+++ NEWS.md
@@ -17,9 +17,9 @@
 
 - New `sqliteVersion()` prints the header and library versions of RSQLite.
 
-- The `SQLiteResult` method of `dbGetInfo()` has been deprecated: it's now
-  better to access the metadata with individual functions. `dbColumnInfo()`
-  will now return information even before you've retrieved any data.
+- All `dbGetInfo()` methods have been deprecated: it's now better to access the 
+  metadata with individual functions. `dbColumnInfo()` will now return 
+  information even before you've retrieved any data.
 
 - `dbListResults()` is no longer supported. Adding a result registry in the
   connection would add significant work, especially when trying to avoid